fix/address-master

+++ src/main/java/com/asg/common/services/service/CommonDataService.java
@@ -10,6 +10,7 @@ import com.asg.common.services.dto.AddressDetailsListResponseDto;
 import com.asg.common.services.dto.AddressPoidResponseDto;
 import com.asg.common.services.dto.CurrencyRateResponseDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CommonDataService {
@@ -67,7 +68,7 @@ public interface CommonDataService {
 
     AddressDetailsListResponseDto fetchAddressByMasterPoid(
             Long addressMasterPoid,
-            Long addressPoid
+            BigDecimal addressPoid
     );
 
 }
